transform to static lib

+++ Niking2D/src/Niking2D/ImGui/ImGuiLayer.cpp
@@ -98,6 +98,9 @@ namespace Niking2D {
 		static bool show = true;
 		ImGui::ShowDemoWindow(&show);
 
+		//ImGui::Begin("");
+		//ImGui::End();
+
 	}
 
 
@@ -248,7 +251,7 @@ namespace Niking2D {
 	bool ImGuiLayer::OnWindowResizeEvent(WindowResizeEvent & e)
 	{
 		ImGuiIO&io = ImGui::GetIO();
-		io.DisplaySize = ImVec2(e.GetWidth(), e.GetHeight());
+		io.DisplaySize = ImVec2((float)e.GetWidth(), (float)e.GetHeight());
 		io.DisplayFramebufferScale = ImVec2(1.0f, 1.0f);
 		glViewport(0, 0, e.GetWidth(), e.GetHeight());
